refactor: minor loss refactoring

+++ tumor_semseg/loss/dice_ce_edge_loss.py
@@ -22,7 +22,7 @@ class DiceLoss(nn.Module):
         inputs = F.softmax(inputs, dim=1)
         inputs = inputs.view(inputs.size(0), self.n_classes, -1)
 
-        targets_one_hot = F.one_hot(targets.long(), self.n_classes).permute(0, 3, 1, 2).float()
+        targets_one_hot = F.one_hot(targets, self.n_classes).permute(0, 3, 1, 2).float()
         targets_one_hot = targets_one_hot.view(targets_one_hot.size(0), self.n_classes, -1)
 
         intersection = (inputs * targets_one_hot).sum(-1)
@@ -45,7 +45,7 @@ class EdgeLoss(nn.Module):
         self.sobel_y.weight = nn.Parameter(sobel_kernel_y, requires_grad=False)
 
     def forward(self, inputs, targets):
-        targets_one_hot = F.one_hot(targets.long(), self.n_classes).permute(0, 3, 1, 2).float()
+        targets_one_hot = F.one_hot(targets, self.n_classes).permute(0, 3, 1, 2).float()
         loss = 0.0
         for class_idx in range(self.n_classes):
             class_input = inputs[:, class_idx : class_idx + 1, :, :]
@@ -78,9 +78,9 @@ class DiceCEEdgeLoss(nn.Module):
         self.edge_loss = EdgeLoss(config.n_classes, config.edge_weight)
 
     def forward(self, inputs, targets):
-        targets = targets.squeeze(1)
+        targets = targets.squeeze(1).long()
         dice_loss = self.dice_loss(inputs, targets)
-        ce_loss = F.cross_entropy(inputs, targets.long(), self.config.ce_weight)
+        ce_loss = F.cross_entropy(inputs, targets, self.config.ce_weight)
         edge_loss = self.edge_loss(inputs, targets)
         total_loss = self.config.alpha * dice_loss + self.config.beta * ce_loss + self.config.gamma * edge_loss
